feat: add ts, eslint, prettier and jest
