+++ .eslintrc.json
@@ -1,0 +1,29 @@
+{
+  "env": {
+    "es2021": true,
+    "node": true,
+    "jest/globals": true
+  },
+  "extends": ["standard-with-typescript", "prettier"],
+  "plugins": ["jest", "prettier"],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "rules": {
+    "@typescript-eslint/strict-boolean-expressions": "off",
+    "@typescript-eslint/naming-convention": "off",
+    "prettier/prettier": [
+      "error",
+      {
+        "semi": true,
+        "tabWidth": 2,
+        "printWidth": 140,
+        "singleQuote": true,
+        "bracketSameLine": false,
+        "trailingComma": "all",
+        "arrowParens": "always"
+      }
+    ]
+  }
+}
