+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+  "semi": true,
+  "printWidth": 140,
+  "tabWidth": 2,
+  "singleQuote": true,
+  "bracketSpacing": true,
+  "trailingComma": "all",
+  "arrowParens": "always"
+}
