--- index.js
@@ -1,8 +1,0 @@
-import { httpServer } from "./src/http_server/index.js";
-import _wss from "./src/ws_server/index.js";
-
-const HTTP_PORT = 8181;
-
-console.log(`Start static http server on the ${HTTP_PORT} port!`);
-httpServer.listen(HTTP_PORT);
-
