+++ jest.config.js
@@ -1,0 +1,24 @@
+const config = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  testPathIgnorePatterns: ['/node_modules/'],
+  testMatch: ['**/?(*.)+(test).ts'],
+  moduleFileExtensions: ['ts', 'js', 'json'],
+  extensionsToTreatAsEsm: ['.ts'],
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+  },
+  restoreMocks: true,
+  resetMocks: true,
+  moduleDirectories: ['node_modules', '<rootDir>/src'],
+  transform: {
+    '^.+\\.(mt|t|cj|j)s$': [
+      'ts-jest',
+      {
+        useESM: true,
+      },
+    ],
+  },
+};
+
+export default config;
