+++ nodemon.json
@@ -1,0 +1,8 @@
+{
+  "watch": ["src"],
+  "ext": "ts",
+  "ignore": ["src/**/*.spec.ts"],
+  "execMap": {
+    "ts": "node --no-warnings=ExperimentalWarning --loader ts-node/esm"
+  }
+}
