+++ package.json
@@ -5,15 +5,40 @@
     "main": "index.js",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start:dev": "nodemon ./index.ts",
+        "start:prod": "node ./build/index.js",
+        "build": "tsc && tsc-alias",
+        "lint": "eslint ./src/**/*.ts --fix",
+        "format": "prettier --write ./**/*.{ts,json}",
+        "test": "NODE_OPTIONS=--experimental-vm-modules jest --coverage=true -w=1 --forceExit --detectOpenHandles --watchAll=false --testPathPattern=src/__tests__ --testPathIgnorePatterns=build"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/jest": "^29.5.14",
+        "@types/node": "^22.15.18",
+        "@types/ws": "^8.18.1",
+        "@typescript-eslint/eslint-plugin": "^6.21.0",
+        "cross-env": "^7.0.3",
+        "eslint": "^8.57.1",
+        "eslint-config-prettier": "^10.1.5",
+        "eslint-config-standard-with-typescript": "^43.0.1",
+        "eslint-plugin-import": "^2.31.0",
+        "eslint-plugin-jest": "^28.11.0",
+        "eslint-plugin-n": "^16.6.2",
+        "eslint-plugin-prettier": "^5.4.0",
+        "eslint-plugin-promise": "^6.6.0",
+        "jest": "^29.7.0",
+        "nodemon": "^2.0.16",
+        "prettier": "^3.5.3",
+        "ts-jest": "^29.3.3",
+        "ts-node": "^10.9.2",
+        "tsc-alias": "^1.8.16",
+        "tsconfig-paths": "^4.2.0",
+        "tsx": "^4.19.4",
+        "typescript": "^5.8.3",
+        "typescript-transform-paths": "^3.5.5"
     },
     "dependencies": {
         "ws": "^8.8.0"
