+++ src/ws_server/index.ts
@@ -1,6 +1,4 @@
-import { WebSocketServer } from 'ws';
-
-const wss = new WebSocketServer({ port: 3000 });
+import wss from 'index.ts';
 
 wss.on('connection', function connection(ws) {
   ws.on('error', console.error);
@@ -11,5 +9,3 @@ wss.on('connection', function connection(ws) {
 
   ws.send(JSON.stringify('something'));
 });
-
-export default wss;
